+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added extra reserved named _(see pull request for list of names)_, pr #448.
 * Use `arc4random_buf` when compiling for Apple, pr #449.
 * Fixed: timeout joining room after rename collection, issue #450.
+* Fixed: return correct wrapped type after rename type, issue #451.
 
 # v1.8.8